formatting

+++ assignment-section-05/classification_model/processing/data_manager.py
@@ -23,31 +23,31 @@ def get_first_cabin(row):
 def get_title(passenger: str) -> str:
     """Extracts the title (Mr, Ms, etc) from the name variable."""
     line = passenger
-    if re.search('Mrs', line):
-        return 'Mrs'
-    elif re.search('Mr', line):
-        return 'Mr'
-    elif re.search('Miss', line):
-        return 'Miss'
-    elif re.search('Master', line):
-        return 'Master'
+    if re.search("Mrs", line):
+        return "Mrs"
+    elif re.search("Mr", line):
+        return "Mr"
+    elif re.search("Miss", line):
+        return "Miss"
+    elif re.search("Master", line):
+        return "Master"
     else:
-        return 'Other'
+        return "Other"
 
 
 def pre_pipeline_preparation(*, dataframe: pd.DataFrame) -> pd.DataFrame:
     # replace question marks with NaN values
-    data = dataframe.replace('?', np.nan)
+    data = dataframe.replace("?", np.nan)
 
     # retain only the first cabin if more than
     # 1 are available per passenger
-    data['cabin'] = data['cabin'].apply(get_first_cabin)
+    data["cabin"] = data["cabin"].apply(get_first_cabin)
 
-    data['title'] = data['name'].apply(get_title)
+    data["title"] = data["name"].apply(get_title)
 
     # cast numerical variables as floats
-    data['fare'] = data['fare'].astype('float')
-    data['age'] = data['age'].astype('float')
+    data["fare"] = data["fare"].astype("float")
+    data["age"] = data["age"].astype("float")
 
     # drop unnecessary variables
     data.drop(labels=config.model_config.unused_fields, axis=1, inplace=True)
